ReactNative project skeleton
This PR introduces the project skeleton for our ReactNative
mobile application.  It contains just a single view, and the
bare minumum that is needed to run the app in simulators or on
your mobile device.  It also has some nice developer features
such as live-reload & es6/7 support.

Summary of changes:
- Basic project skeleton for ReactNative app
- Updated the README with basic instructions to install & run the app
- Included some helpful hints for working with ReactNative & Expo in REACTNATIVE.md